- Implemented rating on both modified and original recipes ( there's still an error but only in yakisoba's recipe)

+++ app/src/main/java/com/letsgetcactus/cocinaconcatalina/ui/screens/ListRecipeHostScreen.kt
@@ -182,7 +182,7 @@ private fun RecipeCard(
                     color = MaterialTheme.colorScheme.onSecondary
                 )
                 RecipeRating(
-                    recipe.avgRating,
+                    recipe.avgRating.toInt(),
                     difficulty = recipe.dificulty ?: DificultyEnum.EASY
                 )
             }
